requirements.txt
discord.py
gradio
